+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- 0.2.1
+  - improving SettingsParser
+  - creating entry [LOCAL][home] in settings for problems directory
 - 0.2.0
   - use up_down diff auto if term size is low
   - rename from solver. to draft.